requirements.txt
pip
Flask
cassandra-driver
requests
json2html
requests_cache
